+++ docs/changes.rst
@@ -58,6 +58,9 @@ Released: not yet
 
 * Mock support: Fixes for storage groups and added support for storage volumes.
 
+* Mock support: Fixed that operations on activation profiles succeed with an
+  empty result set in case the CPC is in DPM mode, instead of failing.
+
 **Enhancements:**
 
 * Added support for Python 3.10. This required increasing the minimum version of